Create zone1
In zone 1 (directly in front of the robot, some 30 inches away) are 3 cubes, 1 of which is an aid pack (to pick up) and the other 2 are mines (to avoid). The cubes are about 18 inches away from each other. This code is supposed to (haven't tested) find the aid pack and bring it back to the "safe zone".